Refactor: pull out visualizer from bay_hopt
