+++ src/pipeline_optimization/bayesian_hopt.py
@@ -1,8 +1,6 @@
 import numpy as np
 import pandas as pd
 
-import plotly.graph_objects as go
-
 from functools import partial
 from hyperopt import hp, tpe, fmin, STATUS_OK, Trials, rand
 from hyperopt.fmin import generate_trials_to_calculate
@@ -153,65 +151,3 @@ class BayesianHopt:
         self.results = results
 
         return results
-
-    def visualize_search_performance(
-        self, xaxis="iterations", all_losses=False, crossvalidation=False
-    ):
-        results = self.results
-
-        # create figure
-        fig = go.Figure()
-
-        # define x-axis
-        if xaxis == "iterations":
-            idx = results.index
-        if xaxis == "walltime":
-            idx = pd.Series(
-                [
-                    results.loc[:i, ("results", "walltime")].sum()
-                    for i in range(len(results))
-                ]
-            )
-
-        # compute best-so-far series
-        rolling_min = pd.Series(
-            [results.loc[:i, ("results", "loss")].min() for i in range(len(results))]
-        )
-        fig.add_trace(
-            go.Scatter(x=idx, y=rolling_min, mode="lines", name="Best so far")
-        )
-
-        # optionally show all computed evaluations
-        if all_losses:
-            loss = results.loc[:, ("results", "loss")]
-            fig.add_trace(
-                go.Scatter(x=idx, y=loss, mode="markers", name="Iteration result")
-            )
-
-        # optionally show the crossvalidation performance
-        if crossvalidation:
-            rolling_crossval = pd.Series(
-                [
-                    results.loc[:i, ("results", "crossval")].min()
-                    for i in range(len(results))
-                ]
-            )
-            fig.add_trace(
-                go.Scatter(
-                    x=idx,
-                    y=rolling_crossval,
-                    mode="lines",
-                    name="Crossvalidation so far",
-                )
-            )
-
-        # optionally show all computed evaluations of crossvalidation performance
-        if all_losses & crossvalidation:
-            crossval = results.loc[:, ("results", "crossval")]
-            fig.add_trace(
-                go.Scatter(
-                    x=idx, y=crossval, mode="markers", name="Crossvalidation result"
-                )
-            )
-
-        fig.show()
